Review changes #2: bye bye node api's
