+++ src/events/commands.ts
@@ -1,8 +1,8 @@
 import { Events, type Message, TextChannel } from "discord.js";
 import { commands } from "../collectCommands.ts";
-import type { BotEvent } from "../customTypes.ts";
+import type { BotEvent } from "../types.ts";
 
-export const nonSlashCommandEvent: BotEvent = {
+export const commandEvent: BotEvent = {
   type: Events.MessageCreate,
   execute: (message: Message): void => {
     if (!(message.channel instanceof TextChannel)) return;
